--- src/main/java/com/planup/planup/domain/notification/service/NotificationService.java
@@ -1,31 +1,0 @@
-package com.planup.planup.domain.notification.service;
-
-import com.planup.planup.domain.notification.dto.NotificationResponseDTO;
-import com.planup.planup.domain.notification.entity.Notification;
-import com.planup.planup.domain.notification.entity.NotificationType;
-import com.planup.planup.domain.notification.entity.TargetType;
-import com.planup.planup.domain.user.entity.User;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
-
-public interface NotificationService {
-
-    // 읽지 않은 알림 조회 (정렬 포함)
-    List<NotificationResponseDTO.NotificationDTO> getUnreadNotifications(Long receiverId);
-    List<NotificationResponseDTO.NotificationDTO> getUnreadNotificationsWithType(Long receiverId, String type);
-
-
-    // 전체 알림 조회 (정렬 포함)
-    List<NotificationResponseDTO.NotificationDTO> getAllNotifications(Long receiverId);
-
-    // 읽음 처리
-    void markAsRead(Long notificationId, Long userId);
-
-    List<NotificationResponseDTO.NotificationDTO> getTop5RecentByUser(Long userId);
-
-    Notification createNotification(Long receiverId, Long senderId, NotificationType notificationType, TargetType targetType, Long targetId);
-
-    @Transactional
-    Notification createNotification(Long receiverId, Long senderId, NotificationType notificationType, TargetType targetType, Long targetId, List<String> updatedParts);
-}
