+++ src/main/java/com/planup/planup/domain/notification/service/NotificationServiceWrite.java
@@ -11,4 +11,6 @@ public interface NotificationServiceWrite {
     Notification createNotification(Long receiverId, Long senderId, NotificationType notificationType, TargetType targetType, Long targetId);
 
     Notification createNotification(Long receiverId, Long senderId, NotificationType notificationType, TargetType targetType, Long targetId, List<String> updatedParts);
+
+    void markAsRead(Long notificationId, Long userId);
 }
