+++ src/main/java/com/planup/planup/domain/notification/service/NotificationServiceWriteImpl.java
@@ -1,13 +1,15 @@
 package com.planup.planup.domain.notification.service;
 
+import com.planup.planup.apiPayload.code.status.ErrorStatus;
+import com.planup.planup.apiPayload.exception.custom.NotificationError;
 import com.planup.planup.domain.notification.entity.Notification;
 import com.planup.planup.domain.notification.entity.NotificationType;
 import com.planup.planup.domain.notification.entity.TargetType;
 import com.planup.planup.domain.notification.entity.device.NotificationCreatedEvent;
 import com.planup.planup.domain.notification.repository.NotificationRepository;
 import com.planup.planup.domain.user.entity.User;
-import com.planup.planup.domain.user.service.UserService;
 import com.planup.planup.domain.user.service.query.UserQueryService;
+import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
@@ -72,5 +74,23 @@ public class NotificationServiceWriteImpl implements NotificationServiceWrite {
                 ));
 
         return savedNotification;
+    }
+
+    @Override
+    public void markAsRead(Long notificationId, Long userId) {
+        Notification notification = notificationRepository.findById(notificationId)
+                .orElseThrow(() -> new EntityNotFoundException("Notification not found"));
+
+        validateReceiver(notification, userId);
+
+        notification.markAsRead(true);
+    }
+
+    //해당 변경을 할 수 있는 권한이 있는지 확인
+    private void validateReceiver(Notification notification, Long userId) {
+        User user = userService.getUserByUserId(userId);
+        if (!notification.getReceiver().equals(user)) {
+            throw new NotificationError(ErrorStatus.UNAUTHORIZED_NOTIFICATION_ACCESS);
+        }
     }
 }
