+++ src/main/java/com/planup/planup/domain/report/service/WeeklyReportService/WeeklyReportServiceImpl.java
@@ -9,7 +9,6 @@ import com.planup.planup.domain.global.message.MessageResponse;
 import com.planup.planup.domain.goal.entity.mapping.UserGoal;
 import com.planup.planup.domain.goal.service.UserGoalService;
 import com.planup.planup.domain.notification.dto.NotificationResponseDTO;
-import com.planup.planup.domain.notification.service.NotificationService;
 import com.planup.planup.domain.report.converter.DailyRecordConverter;
 import com.planup.planup.domain.report.converter.WeeklyReportResponseConverter;
 import com.planup.planup.domain.report.dto.WeeklyReportResponseDTO;
